refactor: added syncCookies getter for the session
