+++ src/proxy/index.ts
@@ -22,7 +22,6 @@ import BUILTIN_HEADERS from '../request-pipeline/builtin-header-names';
 import logger from '../utils/logger';
 import errToString from '../utils/err-to-string';
 import { parse as parseJSON } from '../utils/json';
-import { formatSyncCookie } from '../utils/cookie';
 
 const SESSION_IS_NOT_OPENED_ERR = 'Session is not opened in proxy';
 
@@ -159,31 +158,7 @@ export default class Proxy extends Router {
 
                 logger.serviceMsg.onMessage(msg, result);
 
-                const syncCookies       = session.cookies.syncCookies;
-                const parsedSyncCookies = [] as string[];
-
-                while (true) {
-                    const syncCookie = syncCookies.pop();
-
-                    if (!syncCookie)
-                        break;
-
-                    const cookieRecord = {
-                        ...syncCookie,
-                        sid:          msg.sessionId,
-                        isServerSync: true,
-                        domain:       syncCookie.domain || '',
-                        path:         syncCookie.path || '',
-                        lastAccessed: new Date(),
-                        syncKey:      ''
-                    };
-
-                    const parsedSyncCookie = formatSyncCookie(cookieRecord);
-
-                    parsedSyncCookies.push(parsedSyncCookie);
-                }
-
-                res.setHeader('Set-Cookie', parsedSyncCookies);
+                res.setHeader('Set-Cookie', session.syncCookies);
 
                 respondWithJSON(res, result, false);
             }
